rtsf-at-checkout-cv-region-of-interest/requirements.txt
# Copyright © 2022 Intel Corporation. All rights reserved.
# SPDX-License-Identifier: BSD-3-Clause

paho_mqtt==1.6.1
requests==2.32.0
opencv-python-headless==4.7.0.72
tensorflow==2.11.1
tensorflow-serving-api==2.11.1
grpcio==1.53.2